Fix vote parse error

+++ app/src/main/java/me/ykrank/s1next/data/api/app/model/AppVote.kt
@@ -53,7 +53,7 @@ class AppVote {
      * 过期时间
      */
     @JsonProperty("expiration")
-    var expiration: Int = 0
+    var expiration: Long = 0
     /**
      * 投票的人数
      */
